Compact, Barrel nodes
