+++ projects/atft/src/lib/actor/data-center/atft-data-center-actor.module.ts
@@ -45,6 +45,8 @@ import {DagreCompositionComponent, DagreEdgeComponent, DagreLayoutComponent, Dag
   ],
   entryComponents: [
     ServerStandActorComponent,
+    ServerBarrelActorComponent,
+    ServerCompactActorComponent,
     DagreNodeComponent,
     DagreEdgeComponent,
     DagreCompositionComponent
