+++ projects/atft/src/lib/actor/data-center/layout/dagre-model.ts
@@ -1,0 +1,28 @@
+import * as dagre from 'dagre';
+
+export interface BaseInfo {
+  name: string;
+  label?: string;
+  composition?: string;
+}
+
+export interface Node extends BaseInfo {
+  type?: string;
+  icon?: string;
+}
+
+export interface Composition extends BaseInfo {
+  border?: string;
+}
+
+export interface Edge extends BaseInfo {
+  from: string;
+  to: string;
+}
+
+export interface GraphModel {
+  layout?: dagre.GraphLabel;
+  compositions?: Array<Composition>;
+  nodes?: Array<Node>;
+  edges?: Array<Edge>;
+}
