+++ projects/atft/src/lib/actor/data-center/layout/dagre-yaml-parser.component.ts
@@ -2,9 +2,11 @@ import {
   AfterViewInit,
   Component,
   ComponentFactoryResolver,
+  EventEmitter,
   Input,
   OnChanges,
   Optional,
+  Output,
   SimpleChanges,
   SkipSelf,
   ViewChild,
@@ -14,11 +16,12 @@ import {AbstractEmptyDirective, AbstractObject3D} from '../../../object';
 import {RendererService} from '../../../renderer';
 import {provideParent} from '../../../util';
 import * as yaml from 'yaml';
-import {Composition, Edge, GraphModel, Node} from './dagre-utils';
-import {ServerStandActorComponent} from '../server';
+import {Composition, Edge, GraphModel, Node} from './dagre-model';
+import {ServerBarrelActorComponent, ServerCompactActorComponent, ServerStandActorComponent} from '../server';
 import {DagreNodeComponent} from './dagre-node.component';
 import {DagreEdgeComponent} from './dagre-edge.component';
 import {DagreCompositionComponent} from './dagre-composition.component';
+
 
 function onlyUnique(value, index, self) {
   return self.indexOf(value) === index;
@@ -33,6 +36,7 @@ function onlyUnique(value, index, self) {
 export class DagreYamlParserComponent extends AbstractEmptyDirective implements OnChanges, AfterViewInit {
 
   @Input() yaml;
+  @Output() status = new EventEmitter<boolean>();
 
   @ViewChild('container', {read: ViewContainerRef}) container;
 
@@ -66,15 +70,32 @@ export class DagreYamlParserComponent extends AbstractEmptyDirective implements 
 
   public parseAndCreate() {
     // console.log('DagreYamlParserComponent.parseAndCreate');
-    this.destroyAll();
-    const model: GraphModel = yaml.parse(this.yaml);
+    try {
+      this.destroyAll();
+      const model: GraphModel = yaml.parse(this.yaml);
 
-    // console.log('DagreYamlParserComponent.parseAndCreate yaml', model);
-    if (model && model.nodes && model.nodes.length > 0) {
+      // console.log('DagreYamlParserComponent.parseAndCreate yaml', model);
+      if (model && model.nodes && model.nodes.length > 0) {
 
-      model.compositions?.forEach(i => this.createComposition(i));
-      model.nodes?.forEach(i => this.createNode(i));
-      model.edges?.forEach(i => this.createEdge(i));
+        model.compositions?.forEach(i => this.createComposition(i));
+        model.nodes?.forEach(i => this.createNode(i));
+        model.edges?.forEach(i => this.createEdge(i));
+      }
+      this.status.emit(true);
+    } catch (e) {
+      console.warn('DagreYamlParserComponent.parseAndCreate failed', e);
+      this.status.emit(false);
+      throw e;
+    }
+  }
+
+  protected getNodeComponent(type: string) {
+    if (type === 'compact') {
+      return this.resolver.resolveComponentFactory(ServerCompactActorComponent);
+    } else if (type === 'barrel') {
+      return this.resolver.resolveComponentFactory(ServerBarrelActorComponent);
+    } else {
+      return this.resolver.resolveComponentFactory(ServerStandActorComponent);
     }
   }
 
@@ -86,7 +107,7 @@ export class DagreYamlParserComponent extends AbstractEmptyDirective implements 
     nodeRef.instance.composition = node.composition;
     this.instances.push(nodeRef);
 
-    const serverFactory = this.resolver.resolveComponentFactory(ServerStandActorComponent);
+    const serverFactory = this.getNodeComponent(node.type);
     const serverRef = nodeRef.instance.container.createComponent(serverFactory);
     serverRef.instance.name = node.name;
     serverRef.instance.label = (node.label ? node.label : node.name);
