+++ projects/atft/src/lib/actor/data-center/layout/index.ts
@@ -3,3 +3,4 @@ export * from './dagre-node.component';
 export * from './dagre-edge.component';
 export * from './dagre-composition.component';
 export * from './dagre-yaml-parser.component';
+export * from './dagre-model';
