+++ src/stories/dagre-layout/dagre-sample.stories.ts
@@ -53,7 +53,7 @@ class StorybookDagreComponent {
 
       <atft-dagre-composition name="presentation" label="Presentation tier"></atft-dagre-composition>
       <atft-dagre-composition name="application" label="Application tier"></atft-dagre-composition>
-      <atft-dagre-composition name="data" label="Data tier"></atft-dagre-composition>
+      <atft-dagre-composition name="data" label="Data tier" border="frame"></atft-dagre-composition>
 
       <atft-dagre-node name="spa" composition="presentation">
         <atft-server-compact-actor label="spa"></atft-server-compact-actor>
